Remove unused awssdk bom
Remove the

```
      <groupId>software.amazon.awssdk</groupId>
      <artifactId>bom</artifactId>
```

dependency.

It has not effect because it is not in the `<dependencyManagement>`
section but in the `<dependencies>` section.

There is no need for the bom because mod-bulk-operations doesn't
directly use any awssdk library.

folio-s3-client comes with a recent version of awssdk, no need to
overwrite that version.